+++ examples/widgets.rs
@@ -7,8 +7,6 @@ enum Action {
     AddItem,
     ClearText,
     EntryActivated(Entity),
-    EntryChanged(Entity),
-    ValueChanged(Entity),
     IncrementCounter,
     RemoveItem,
     ToggleTheme(Entity),
@@ -86,12 +84,6 @@ impl State for MainViewState {
                     let text = text_box.text_mut();
                     println!("submitting {}", text);
                     text.clear();
-                }
-                Action::EntryChanged(entity) => {
-                    println!("entry changed: {}", text_box(ctx.get_widget(entity)).text());
-                }
-                Action::ValueChanged(_entity) => {
-                    //println!("Slider value changed");
                 }
                 Action::ToggleTheme(entity) => {
                     let light = *ctx.get_widget(entity).get::<bool>("selected");
@@ -276,9 +268,6 @@ impl Template for MainView {
                                     .on_activate(move |states, entity| {
                                         state(id, states).action(Action::EntryActivated(entity));
                                     })
-                                    // .on_changed(move |states, entity| {
-                                    //     state(id, states).action(Action::EntryChanged(entity));
-                                    // })
                                     .build(ctx),
                             )
                             .child(
@@ -291,9 +280,6 @@ impl Template for MainView {
                                     .on_activate(move |states, entity| {
                                         state(id, states).action(Action::EntryActivated(entity));
                                     })
-                                    // .on_changed(move |states, entity| {
-                                    //     state(id, states).action(Action::EntryChanged(entity));
-                                    // })
                                     .build(ctx),
                             )
                             .child(
@@ -462,9 +448,11 @@ impl Template for MainView {
                             )
                             .child(
                                 Switch::new()
-                                    // .on_changed(move |states, entity| {
-                                    //     state(id, states).action(Action::ToggleTheme(entity));
-                                    // })
+                                    .on_changed(move |states, entity, key| {
+                                        if key == "selected" {
+                                            state(id, states).action(Action::ToggleTheme(entity));
+                                        }
+                                    })
                                     .v_align("center")
                                     .build(ctx),
                             )