+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Improve: Mouse event arguments
 * Fix: Crash when a child widget is removed
 * TabWidget widget
+* Add on_changed property change callback to all widgets
 
 ### 0.3.1-alpha2